fixed alle the polymorphic typing ... next, replace everything
