+++ packages/types/src/react.ts
@@ -1,101 +1,94 @@
-import React from 'react';
-
-// Simple merge
-type Merge<T, U> = Omit<T, keyof U> & U;
+import * as React from 'react';
 
 /**
- * Get props without supporting `ref` and strip the `style` prop.
- *
- * @example <caption>Will contain all HTML attributes of a regular button element, minus "style"</caption>
- * type ButtonProps = ComponentProps<'button'>
+ * Infer the props of a component `C`.
  */
-export type ComponentProps<T extends React.ElementType> = Omit<
-  React.ComponentPropsWithoutRef<T>,
-  'style'
->;
-
-/**
- * Get props with supporting `ref` and strip the `style` prop.
- *
- * @example <caption>Will contain all HTML attributes of a regular button element, minus "style"</caption>
- * type ButtonProps = ComponentProps<'button'>
- */
-export type ComponentPropsWithRef<T extends React.ElementType> = Omit<
-  React.ComponentPropsWithRef<T>,
-  'style'
->;
+export type PropsOf<C> = C extends React.FC<infer P>
+  ? P
+  : C extends React.Component<infer Props>
+  ? Props
+  : never;
 
 /**********************************************/
 /*                                            */
 /*            POLYMORPIC COMPONENT            */
 /*                                            */
 /**********************************************/
-
-// Remove call signature from `ForwardRefExoticComponent` so we can override it
-type ForwardRefComponent<P> = Pick<
-  React.ForwardRefExoticComponent<P>,
-  keyof React.ForwardRefExoticComponent<any>
->;
-
 /**
- * Props for the polymorphic `as` prop.
+ * Polymorphic types are based on Radix's (now deprecated) types.
+ * They can be found here: https://github.com/radix-ui/primitives/blob/2f139a832ba0cdfd445c937ebf63c2e79e0ef7ed/packages/react/polymorphic/src/polymorphic.ts
  */
-export type PropsWithAs<P, T extends React.ElementType> = P & { as?: T };
 
 /**
- * Helper type to create props including the polymorphic `as` prop, but no `ref`.
- */
-export type PolymorphicProps<P, T extends React.ElementType> = Merge<
-  ComponentProps<T>,
-  PropsWithAs<P, T>
->;
-
-/**
- * Helper type to create props including the polymorphic `as` prop and forwarding `ref`.
- */
-export type PolymorphicPropsWithRef<P, T extends React.ElementType> = Merge<
-  ComponentPropsWithRef<T>,
-  PropsWithAs<P, T>
->;
-
-/**
- * Create a polymorphpic component with the `as` prop.
+ * Merge two objects into one. Second one will override properties of the
+ * first one.
  *
- * This type can be used to create a component that allows to change the rendered
- * element via its `as` prop.
+ * Note: We can not use `Merge` from type-fest here, because for whatever reasons
+ * the optional modifier will be removed.
  *
- * @example
- * const Component: = ((props:ComponentProps) => <div>{props.children}</div>))
- *  as PolymorphicComponent<ComponentProps, 'div'>;
+ * @internal
  */
-export interface PolymorphicComponent<P, T extends React.ElementType>
-  extends React.Component<any> {
-  (
-    props: P & React.ComponentProps<T> & { as?: never }
-  ): React.ReactElement<P> | null;
+type Merge<P1 = {}, P2 = {}> = Omit<P1, keyof P2> & P2;
+
+/**
+ * Used to merge properties of an HTML element with additional props defined in a component.
+ *
+ * @internal
+ */
+type MergeProps<T, P = {}> = P &
+  Merge<T extends React.ElementType ? React.ComponentPropsWithRef<T> : {}, P>;
+
+/**
+ * Narrow down instrinsic elements to the keys (HTML tags).
+ *
+ * @internal
+ */
+type NarrowIntrinsic<E> = E extends keyof JSX.IntrinsicElements ? E : never;
+
+/**
+ * Infer the (additional) props of the component `C` if it a `PolymorphicComponent`.
+ * This not include any properties added to a polymorphic component based on its
+ * instrinstic element type.
+ */
+export type OwnProps<C> = C extends PolymorphicComponent<any, infer P>
+  ? P
+  : never;
+
+export type IntrinsicElement<C> = C extends PolymorphicComponent<infer I, any>
+  ? I
+  : never;
+
+export interface PolymorphicComponent<
+  T,
+  P = {}
+  /**
+   * Extends original type to ensure built in React types play nice
+   * with polymorphic components still e.g. `React.ElementRef` etc.
+   */
+> extends React.ForwardRefExoticComponent<MergeProps<T, P & { as?: T }>> {
+  /**
+   * USE CASE: `as` porp is a string:
+   *
+   * Merges original own props (without DOM props) and the inferred props
+   * from `as` element with the own props taking precendence.
+   *
+   * We explicitly define a `JSX.IntrinsicElements` overload so that
+   * events are typed for consumers.
+   */
+  <As extends keyof JSX.IntrinsicElements = NarrowIntrinsic<T>>(
+    props: MergeProps<As, P & { as: As }>
+  ): React.ReactElement | null;
+
+  /**
+   * USE CASE: `as` porp is a component:
+   *
+   * Merges original own props (without DOM props) and the inferred props
+   * from `as` element with the own props taking precendence.
+   *
+   * We don't use `React.ComponentType` here as we get type errors
+   * when consumers try to do inline `as` components.
+   */
   <As extends React.ElementType>(
-    props: PolymorphicProps<P, As>
-  ): React.ReactElement<P> | null;
-}
-
-/**
- * Create a polymorphpic component with the `as` prop that also support passing a `ref`.
- *
- * This type can be used to create a component that allows to change the rendered
- * element via its `as` prop and forwarding a `ref` to it.
- *
- * **Note: you should use React's `forwardRef` with this type.**
- *
- * @example
- * const Component: PolymorphicComponentWithRef<ComponentProps, 'div'>
- *   = forwardRef((props, ref) => <div>{props.children}</div>);
- */
-export interface PolymorphicComponentWithRef<P, T extends React.ElementType>
-  extends ForwardRefComponent<any> {
-  <As extends React.ElementType>(
-    props: PolymorphicPropsWithRef<P, As>
-  ): React.ReactElement<P> | null;
-  (
-    props: P & React.ComponentPropsWithRef<T> & { as?: never }
-  ): React.ReactElement<P> | null;
+    props: MergeProps<As, P & { as: As }>
+  ): React.ReactElement | null;
 }
